refactor dialWorkerLoop into an object
